refactor: :recycle: seperate components
